# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ok {
  ok: boolean;
}

export type ResultType = (typeof ResultType)[keyof typeof ResultType];

export const ResultType = {
  all_or_nothing: "all_or_nothing",
  stuck_loop: "stuck_loop",
  overwhelmed: "overwhelmed",
  lost_herself: "lost_herself",
} as const;

export type LeadCreateUtm = {
  source?: string;
  medium?: string;
  campaign?: string;
  content?: string;
  term?: string;
} | null;

export interface LeadCreate {
  /**
   * @minLength 1
   * @maxLength 80
   */
  firstName: string;
  email: string;
  phone?: string | null;
  consent: boolean;
  resultType: ResultType;
  sessionId?: string | null;
  source?: string | null;
  referrer?: string | null;
  deviceType?: string | null;
  utm?: LeadCreateUtm;
}

export interface Lead {
  id?: number;
  firstName?: string;
  email?: string;
  phone?: string | null;
  resultType?: ResultType;
  status?: string;
  bookedAt?: string | null;
  unsubscribed?: boolean;
  createdAt?: string;
}

export type FunnelSettingsResultVideos = { [key: string]: string };

export interface FunnelSettings {
  bookingUrl?: string | null;
  publicSiteUrl?: string | null;
  trainingVideoUrl?: string | null;
  resultVideos?: FunnelSettingsResultVideos;
}

export interface BookingIntakeCreate {
  email: string;
  resultType: ResultType;
  biggestStruggle?: string;
  goal90Days?: string;
  triedBefore?: string;
  knockedOff?: string;
  howSoon?: string;
  openToCoaching?: string;
  bestPhone?: string;
  bestTime?: string;
}

export type AdminStatsTotals = {
  quizStarts?: number;
  quizCompletions?: number;
  leadsCaptured?: number;
  bookCtaClicks?: number;
  trainingClicks?: number;
  trainingViews?: number;
  intakeCompleted?: number;
  calendarOpened?: number;
  bookedCalls?: number;
  resultViewed?: number;
  conversionRate?: number;
};

export type AdminStatsBreakdownItem = {
  resultType?: ResultType;
  count?: number;
};

export interface AdminStats {
  totals: AdminStatsTotals;
  breakdown: AdminStatsBreakdownItem[];
}

export type AdminLeadDetailSubmissionScoreJson = { [key: string]: number };

export type AdminLeadDetailSubmissionAnswersJson = { [key: string]: string };

export type AdminLeadDetailSubmission = {
  id?: number;
  sessionId?: string;
  resultType?: ResultType;
  scoreJson?: AdminLeadDetailSubmissionScoreJson;
  answersJson?: AdminLeadDetailSubmissionAnswersJson;
  completedAt?: string | null;
  [key: string]: unknown;
} | null;

export type AdminLeadDetailEventsItemPayload = {
  [key: string]: unknown;
} | null;

export type AdminLeadDetailEventsItem = {
  id?: number;
  eventName?: string;
  payload?: AdminLeadDetailEventsItemPayload;
  createdAt?: string;
  [key: string]: unknown;
};

export type AdminLeadDetailIntakesItem = { [key: string]: unknown };

export type AdminLeadDetailSequencesItem = {
  id?: number;
  sequenceType?: string;
  currentStep?: number;
  totalSteps?: number;
  status?: string;
  [key: string]: unknown;
};

export type AdminLeadDetailEmailsItem = {
  id?: number;
  subject?: string;
  status?: string;
  sentAt?: string;
  [key: string]: unknown;
};

export interface AdminLeadDetail {
  lead: Lead;
  submission?: AdminLeadDetailSubmission;
  events: AdminLeadDetailEventsItem[];
  intakes: AdminLeadDetailIntakesItem[];
  sequences: AdminLeadDetailSequencesItem[];
  emails: AdminLeadDetailEmailsItem[];
}

/**
 * Raw key→value overrides stored in funnel_settings.
 */
export type AdminSettingsOverrides = { [key: string]: string };

export type AdminSettingsEffectiveResultVideos = { [key: string]: string };

export type AdminSettingsEffective = {
  bookingUrl?: string;
  publicSiteUrl?: string;
  trainingVideoUrl?: string;
  ownerEmail?: string;
  mailingAddress?: string;
  resultVideos?: AdminSettingsEffectiveResultVideos;
};

export interface AdminSettings {
  /** Raw key→value overrides stored in funnel_settings. */
  overrides: AdminSettingsOverrides;
  effective: AdminSettingsEffective;
}

export type QuizStartBody = {
  sessionId: string;
};

export type QuizAnswerBody = {
  sessionId: string;
  questionKey: string;
  answerKey: string;
  answerType: ResultType;
};

/**
 * Map of resultType → numeric score.
 */
export type QuizCompleteBodyScoreJson = { [key: string]: number };

/**
 * Map of questionKey → answerKey selected by the user.
 */
export type QuizCompleteBodyAnswersJson = { [key: string]: string };

export type QuizCompleteBody = {
  sessionId: string;
  resultType: ResultType;
  /** Map of resultType → numeric score. */
  scoreJson: QuizCompleteBodyScoreJson;
  /** Map of questionKey → answerKey selected by the user. */
  answersJson: QuizCompleteBodyAnswersJson;
};

export type SendTestEmailBody = {
  to: string;
  resultType?: ResultType;
};

export type SendTestEmail200 = {
  ok?: boolean;
  [key: string]: unknown;
};

export type CreateLead200 = Ok & {
  leadId?: number;
};

export type TrackEventBodyPayload = { [key: string]: unknown };

export type TrackEventBody = {
  eventName: string;
  payload?: TrackEventBodyPayload;
  sessionId?: string;
  leadId?: number;
};

export type BookingWebhookBody = {
  email?: string;
  [key: string]: unknown;
};

export type BookingWebhook200 = Ok & {
  matched?: boolean;
};

export type UnsubscribeParams = {
  token: string;
};

export type AdminLoginBody = {
  password: string;
};

export type AdminMe200 = {
  authenticated: boolean;
  configured: boolean;
};

export type AdminListLeadsParams = {
  /**
   * @minimum 1
   * @maximum 500
   */
  limit?: number;
};

export type AdminListLeads200 = {
  leads: Lead[];
};

export type AdminPutSettingsBodySettings = { [key: string]: string };

export type AdminPutSettingsBody = {
  settings: AdminPutSettingsBodySettings;
};

export type AdminPutSettings200 = Ok & {
  count?: number;
};