# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/funnel/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Exchange password for an admin session cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string }
              required: [password]
      responses:
        "200":
          description: Authenticated; sets the admin session cookie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "401":
          description: Invalid password
        "503":
          description: Admin not configured
...
  /admin/funnel/leads:
    get:
      operationId: adminListLeads
      tags: [admin]
      summary: Paginated list of captured leads
      parameters:
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: Leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  leads:
                    type: array
                    items: { $ref: "#/components/schemas/Lead" }
                required: [leads]
  /admin/funnel/leads/{id}:
    get:
      operationId: adminLeadDetail
      tags: [admin]
      summary: Lead detail bundle (lead + quiz submission + intakes + sequences + emails + recent events)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminLeadDetail" }
        "404":
          description: Not found
  /admin/funnel/leads/{id}/mark-booked:
    post:
      operationId: adminMarkBooked
      tags: [admin]
      summary: Manually mark a lead as booked (halts their nurture)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
        "404":
          description: Not found
  /admin/funnel/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: Read current funnel settings (env defaults + DB overrides)
      responses:
        "200":
          description: Settings bundle
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSettings" }
    put:
      operationId: adminPutSettings
      tags: [admin]
      summary: Upsert one or more funnel settings; empty string deletes the override
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                settings:
                  type: object
                  additionalProperties: { type: string }
              required: [settings]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Ok"
                  - type: object
                    properties:
                      count: { type: integer }
...
components:
  schemas:
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    ResultType:
      type: string
      enum: [all_or_nothing, stuck_loop, overwhelmed, lost_herself]
...
    Lead:
      type: object
      properties:
        id: { type: integer }
        firstName: { type: string }
        email: { type: string }
        phone: { type: string, nullable: true }
        resultType: { $ref: "#/components/schemas/ResultType" }
        status: { type: string }
        bookedAt: { type: string, format: date-time, nullable: true }
        unsubscribed: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    AdminLeadDetail:
      type: object
      required: [lead, events, intakes, sequences, emails]
      properties:
        lead: { $ref: "#/components/schemas/Lead" }
        submission:
          type: object
          nullable: true
          additionalProperties: true
          properties:
            id: { type: integer }
            sessionId: { type: string }
            resultType: { $ref: "#/components/schemas/ResultType" }
            scoreJson:
              type: object
              additionalProperties: { type: integer }
            answersJson:
              type: object
              additionalProperties: { type: string }
            completedAt: { type: string, format: date-time, nullable: true }
        events:
          type: array
          items:
            type: object
            additionalProperties: true
            properties:
              id: { type: integer }
              eventName: { type: string }
              payload: { type: object, additionalProperties: true, nullable: true }
              createdAt: { type: string, format: date-time }
        intakes:
          type: array
          items:
            type: object
            additionalProperties: true
        sequences:
          type: array
          items:
            type: object
            additionalProperties: true
            properties:
              id: { type: integer }
              sequenceType: { type: string }
              currentStep: { type: integer }
              totalSteps: { type: integer }
              status: { type: string }
        emails:
          type: array
          items:
            type: object
            additionalProperties: true
            properties:
              id: { type: integer }
              subject: { type: string }
              status: { type: string }
              sentAt: { type: string, format: date-time }
    AdminSettings:
      type: object
      required: [overrides, effective]
      properties:
        overrides:
          type: object
          additionalProperties: { type: string }
          description: Raw key→value overrides stored in funnel_settings.
        effective:
          type: object
          properties:
            bookingUrl: { type: string }
            publicSiteUrl: { type: string }
            trainingVideoUrl: { type: string }
            ownerEmail: { type: string }
            mailingAddress: { type: string }
            resultVideos:
              type: object
              additionalProperties: { type: string }