# lib/api-zod/src/generated/types/adminSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminSettingsEffective } from "./adminSettingsEffective";
import type { AdminSettingsOverrides } from "./adminSettingsOverrides";

export interface AdminSettings {
  /** Raw key→value overrides stored in funnel_settings. */
  overrides: AdminSettingsOverrides;
  effective: AdminSettingsEffective;
}